src/config_api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.1
pydantic>=2.11.0
supabase>=2.7.4
python-dotenv>=1.0.1
httpx>=0.28.1
